requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
fastapi==0.110.0
grpcio==1.62.1
grpcio-tools==1.62.1
h11==0.14.0
idna==3.6
protobuf==4.25.3
pydantic==2.6.4
pydantic_core==2.16.3
requests==2.31.0
setuptools==69.1.1
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.28.0
